feat(store): wire obsidian defaults into every content-write route
stage_chunks, stage_summary_with_layout, and write_raw_items all call
ensure_obsidian_defaults_if_enabled so a fresh content root gets the bundled
.obsidian/ graph colour mapping. The wrapper is feature-gated and best-effort
(never aborts persistence over a cosmetic default). Tests relocated to sibling
obsidian_tests.rs and obsidian_registry_tests.rs, with a feature-gated regression
test for the staged graph.json/types.json.
